Add files via upload
Developing a poker playing agent 

Numerous researches have worked on developing software agents which are capable or replicating a humans ability to play strategic games. Poker in particular poses many challenges from an artificial intelligence standpoint due to it being a stochastic, incomplete information game. This thesis describes poker, its rules and the underlying properties and strategies of the game and introduces a number of famous poker agents which at the time of their development, were leading edge solutions. It then moves through the process of architecture design, development and testing of an artificially intelligent poker playing agent. A number of technical development tools were utilised to create the application; these include Java, SQLite and Tesseract. The thesis then describes how the application employs a knowledge based system combining a rule based and formula based approach. A set of pre defined ranges were used for the preflop play and combined with other play information including competitor actions and cards in hand to make decisions based on pre defined rules. A formula based approach is then utilised to calculate the winning chance and make play decisions based on all of the known information. The concluding part of the thesis analyses testing methodologies and overall results of the software agent. Pokergenius was utilised for real world game play testing and against 8 of Pokergeniuses inbuilt players the software agent was able to consistently and accurately make decisions to play poker and in many cases won against a number of the bots. Overall, the application was able to successfully operate as designed.